Split out QUIC related functions
